fix: move the review-item vocabulary out of the layout lane (fixes #413)
