+++ .changeset/review-vocabulary-lane.md
@@ -1,0 +1,5 @@
+---
+'@docx-editor.dev/core': patch
+---
+
+Internal restructuring moves the review-item types and their pure helpers into the store lane; no behavior change.
