+++ docs/api/docx-editor-core/contracts-modules.api.md
@@ -98,7 +98,6 @@ export type ReviewItem = ReviewRevisionItem | ReviewCommentItem | ReviewCustomIt
 export interface ReviewModelInput {
     readonly commentsExtendedPart?: OoxmlPart | undefined;
     readonly commentsPart?: OoxmlPart | undefined;
-    // (undocumented)
     readonly customNodePayloads?: ReadonlyMap<string, {
         readonly nodeId: string;
         readonly label: string;
