+++ packages/core/src/binding/tree-session.ts
@@ -13,8 +13,7 @@
 
 import { projectedText, storyCarriesCommentAnchor } from './story-text-reads.ts';
 import type { Node as PMNode } from 'prosemirror-model';
-import { paragraphOrderOfPart } from '@docx-editor.dev/core/store';
-import type { ReviewItem } from '../layout/review-support.ts';
+import { paragraphOrderOfPart, type ReviewItem } from '@docx-editor.dev/core/store';
 import {
   canApplyLocalReviewPatch,
   localReviewPatchParagraphId,
